requirements.txt
streamlit
fastapi
uvicorn[standard]
requests
python-dotenv
aiofiles
twelvelabs
google-genai
openai
python-multipart
fpdf2
httpx
